refactor: Fix TypeScript build configuration and module ID validation
- Add tsconfig.json for shared/types and shared/module_ids
- Fix module ID fuzzy matching threshold (10 edit distance)
- Fix filter order in findSimilarModules (filter before slice)
- Update test imports to use correct dist paths
- Add comprehensive build and test scripts to root package.json
- All tests passing (11 module_ids tests, 10 MCP server tests)
